Feature: Motion slipperiness

+++ bot/core/AxisAlignedBB.go
@@ -4,10 +4,10 @@ import "math"
 
 type AxisAlignedBB struct {
 	MinX, MinY, MinZ,
-	MaxX, MaxY, MaxZ float32
+	MaxX, MaxY, MaxZ float64
 }
 
-func (a AxisAlignedBB) Contract(x, y, z float32) AxisAlignedBB {
+func (a AxisAlignedBB) Contract(x, y, z float64) AxisAlignedBB {
 	/*
 		Took from net.minecraft.util.math.AxisAlignedBB#contract version 1.12.2
 	*/
@@ -38,7 +38,7 @@ func (a AxisAlignedBB) Contract(x, y, z float32) AxisAlignedBB {
 	return AxisAlignedBB{MinX: d0, MinY: d1, MinZ: d2, MaxX: d3, MaxY: d4, MaxZ: d5}
 }
 
-func (a AxisAlignedBB) Expand(x, y, z float32) AxisAlignedBB {
+func (a AxisAlignedBB) Expand(x, y, z float64) AxisAlignedBB {
 	/*
 		Took from net.minecraft.util.math.AxisAlignedBB#expand version 1.12.2
 	*/
@@ -69,7 +69,7 @@ func (a AxisAlignedBB) Expand(x, y, z float32) AxisAlignedBB {
 	return AxisAlignedBB{MinX: d0, MinY: d1, MinZ: d2, MaxX: d3, MaxY: d4, MaxZ: d5}
 }
 
-func (a AxisAlignedBB) Grow(x, y, z float32) AxisAlignedBB {
+func (a AxisAlignedBB) Grow(x, y, z float64) AxisAlignedBB {
 	/*
 		Took from net.minecraft.util.math.AxisAlignedBB#grow version 1.12.2
 	*/
@@ -86,12 +86,12 @@ func (a AxisAlignedBB) Intersect(other AxisAlignedBB) AxisAlignedBB {
 	/*
 		Took from net.minecraft.util.math.AxisAlignedBB#intersect version 1.12.2
 	*/
-	d0 := float32(math.Max(float64(a.MinX), float64(other.MinX)))
-	d1 := float32(math.Min(float64(a.MinY), float64(other.MinY)))
-	d2 := float32(math.Min(float64(a.MinZ), float64(other.MinZ)))
-	d3 := float32(math.Min(float64(a.MaxX), float64(other.MaxX)))
-	d4 := float32(math.Min(float64(a.MaxY), float64(other.MaxY)))
-	d5 := float32(math.Min(float64(a.MaxZ), float64(other.MaxZ)))
+	d0 := math.Max(a.MinX, other.MinX)
+	d1 := math.Min(a.MinY, other.MinY)
+	d2 := math.Min(a.MinZ, other.MinZ)
+	d3 := math.Min(a.MaxX, other.MaxX)
+	d4 := math.Min(a.MaxY, other.MaxY)
+	d5 := math.Min(a.MaxZ, other.MaxZ)
 	return AxisAlignedBB{MinX: d0, MinY: d1, MinZ: d2, MaxX: d3, MaxY: d4, MaxZ: d5}
 }
